+++ cli/src/cmd/forge/test/filter.rs
@@ -2,8 +2,9 @@ use crate::utils::FoundryPathExt;
 use clap::Parser;
 use ethers::solc::{FileFilter, ProjectPathsConfig};
 use forge::TestFilter;
+use foundry_common::glob::GlobMatcher;
 use foundry_config::Config;
-use std::{fmt, path::Path, str::FromStr};
+use std::{fmt, path::Path};
 
 /// The filter to use during testing.
 ///
@@ -212,69 +213,5 @@ impl TestFilter for ProjectPathsAwareFilter {
 impl fmt::Display for ProjectPathsAwareFilter {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         self.args_filter.fmt(f)
-    }
-}
-
-/// A `globset::Glob` that creates its `globset::GlobMatcher` when its created, so it doesn't need
-/// to be compiled when the filter functions `TestFilter` functions are called.
-#[derive(Debug, Clone)]
-pub struct GlobMatcher {
-    /// The parsed glob
-    pub glob: globset::Glob,
-    /// The compiled `glob`
-    pub matcher: globset::GlobMatcher,
-}
-
-// === impl GlobMatcher ===
-
-impl GlobMatcher {
-    /// Tests whether the given path matches this pattern or not.
-    ///
-    /// The glob `./test/*` won't match absolute paths like `test/Contract.sol`, which is common
-    /// format here, so we also handle this case here
-    pub fn is_match(&self, path: &str) -> bool {
-        let mut matches = self.matcher.is_match(path);
-        if !matches && !path.starts_with("./") && self.as_str().starts_with("./") {
-            matches = self.matcher.is_match(format!("./{path}"));
-        }
-        matches
-    }
-
-    /// Returns the `Glob` string used to compile this matcher.
-    pub fn as_str(&self) -> &str {
-        self.glob.glob()
-    }
-}
-
-impl fmt::Display for GlobMatcher {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.glob.fmt(f)
-    }
-}
-
-impl FromStr for GlobMatcher {
-    type Err = globset::Error;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        s.parse::<globset::Glob>().map(Into::into)
-    }
-}
-
-impl From<globset::Glob> for GlobMatcher {
-    fn from(glob: globset::Glob) -> Self {
-        let matcher = glob.compile_matcher();
-        Self { glob, matcher }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn can_match_glob_paths() {
-        let matcher: GlobMatcher = "./test/*".parse().unwrap();
-        assert!(matcher.is_match("test/Contract.sol"));
-        assert!(matcher.is_match("./test/Contract.sol"));
     }
 }
