+++ cli/tests/fixtures/can_build_skip_glob.stdout
@@ -1,0 +1,3 @@
+Compiling 1 files with 0.8.17
+Solc 0.8.17 finished in 33.25ms
+Compiler run successful!
