+++ common/Cargo.toml
@@ -41,6 +41,7 @@ semver = "1"
 once_cell = "1"
 dunce = "1"
 regex = "1"
+globset = "0.4"
 
 [dev-dependencies]
 tokio = { version = "1", features = ["rt-multi-thread", "macros"] }
