+++ common/src/lib.rs
@@ -11,6 +11,7 @@ pub mod errors;
 pub mod evm;
 pub mod fmt;
 pub mod fs;
+pub mod glob;
 pub mod provider;
 pub mod selectors;
 pub mod shell;
